# Optimize arraySort