--- a/src/Functions/array/arraySort.cpp
+++ b/src/Functions/array/arraySort.cpp
@@ -1,3 +1,7 @@
+#include <Columns/ColumnDecimal.h>
+#include <Columns/ColumnFixedString.h>
+#include <Columns/ColumnString.h>
+#include <Columns/ColumnsDateTime.h>
 #include <Functions/FunctionFactory.h>
 #include <Functions/array/arraySort.h>
 #include <Common/iota.h>
@@ -13,12 +17,72 @@ namespace ErrorCodes
 namespace
 {
 
-template <bool positive>
+template <bool positive, typename ColumnType>
 struct Less
+{
+    const ColumnType & column;
+
+    explicit Less(const IColumn & column_)
+        : column(assert_cast<const ColumnType &>(column_))
+    {
+    }
+
+    bool operator()(size_t lhs, size_t rhs) const
+    {
+        if constexpr (positive)
+            return column.compareAt(lhs, rhs, column, 1) < 0;
+        else
+            return column.compareAt(lhs, rhs, column, -1) > 0;
+    }
+};
+
+template <bool positive, typename ColumnType>
+struct NullableLess
+{
+    const ColumnType & nested_column;
+    const NullMap & null_map;
+
+    explicit NullableLess(const IColumn & nested_column_, const NullMap & null_map_)
+        : nested_column(assert_cast<const ColumnType &>(nested_column_))
+        , null_map(null_map_)
+    {
+    }
+
+    bool operator()(size_t lhs, size_t rhs) const
+    {
+        bool lhs_is_null = null_map[lhs];
+        bool rhs_is_null = null_map[rhs];
+
+        if constexpr (positive)
+        {
+            if (lhs_is_null) [[unlikely]]
+                return false;
+
+            if (rhs_is_null) [[unlikely]]
+                return true;
+
+            return nested_column.compareAt(lhs, rhs, nested_column, 1) < 0;
+        }
+        else
+        {
+            if (rhs_is_null) [[unlikely]]
+                return false;
+
+            if (lhs_is_null) [[unlikely]]
+                return true;
+
+            return nested_column.compareAt(lhs, rhs, nested_column, -1) > 0;
+        }
+    }
+};
+
+
+template <bool positive>
+struct GenericLess
 {
     const IColumn & column;
 
-    explicit Less(const IColumn & column_) : column(column_) { }
+    explicit GenericLess(const IColumn & column_) : column(column_) { }
 
     bool operator()(size_t lhs, size_t rhs) const
     {
@@ -62,23 +126,97 @@ ColumnPtr ArraySortImpl<positive, is_partial>::execute(
     iota(permutation.data(), nested_size, IColumn::Permutation::value_type(0));
 
     ColumnArray::Offset current_offset = 0;
-    for (size_t i = 0; i < size; ++i)
-    {
-        auto next_offset = offsets[i];
-        if constexpr (is_partial)
-        {
-            if (limit)
-            {
-                const auto effective_limit = std::min<size_t>(limit, next_offset - current_offset);
-                ::partial_sort(&permutation[current_offset], &permutation[current_offset + effective_limit], &permutation[next_offset], Less<positive>(*mapped));
-            }
-            else
-                ::sort(&permutation[current_offset], &permutation[next_offset], Less<positive>(*mapped));
-        }
-        else
-            ::sort(&permutation[current_offset], &permutation[next_offset], Less<positive>(*mapped));
-        current_offset = next_offset;
+
+#define APPLY_COMPARATOR(CMP) \
+    for (size_t i = 0; i < size; ++i) \
+    { \
+        auto next_offset = offsets[i]; \
+        if constexpr (is_partial) \
+        { \
+            if (limit) \
+            { \
+                const auto effective_limit = std::min<size_t>(limit, next_offset - current_offset); \
+                ::partial_sort(&permutation[current_offset], &permutation[current_offset + effective_limit], &permutation[next_offset], CMP); \
+            } \
+            else \
+                ::sort(&permutation[current_offset], &permutation[next_offset], CMP); \
+        } \
+        else \
+            ::sort(&permutation[current_offset], &permutation[next_offset], CMP); \
+        current_offset = next_offset; \
     }
+
+#define DISPATCH_FOR_NONNULLABLE_COLUMN(TYPE) \
+    else if (checkAndGetColumn<TYPE>(mapped.get())) \
+    { \
+        Less<positive, TYPE> cmp(*mapped); \
+        APPLY_COMPARATOR(cmp) \
+    }
+
+    if (false)
+        ;
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnUInt8)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnUInt16)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnUInt32)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnUInt64)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnInt8)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnInt16)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnInt32)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnInt64)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnFloat32)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnFloat64)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnDateTime64)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnDecimal<Decimal32>)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnDecimal<Decimal64>)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnDecimal<Decimal128>)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnDecimal<Decimal256>)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnString)
+    DISPATCH_FOR_NONNULLABLE_COLUMN(ColumnFixedString)
+#undef DISPATCH_FOR_NONNULLABLE_COLUMN
+
+    else if (const auto * nullable = checkAndGetColumn<ColumnNullable>(mapped.get()))
+    {
+        const auto & null_map = nullable->getNullMapData();
+
+#define DISPATCH_FOR_NULLABLE_COLUMN(TYPE) \
+    else if (checkAndGetColumn<TYPE>(&nullable->getNestedColumn())) \
+    { \
+        NullableLess<positive, TYPE> cmp(nullable->getNestedColumn(), null_map); \
+        APPLY_COMPARATOR(cmp) \
+    }
+
+        if (false)
+            ;
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnUInt8)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnUInt16)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnUInt32)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnUInt64)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnInt8)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnInt16)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnInt32)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnInt64)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnFloat32)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnFloat64)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnDateTime64)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnDecimal<Decimal32>)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnDecimal<Decimal64>)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnDecimal<Decimal128>)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnDecimal<Decimal256>)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnString)
+        DISPATCH_FOR_NULLABLE_COLUMN(ColumnFixedString)
+        else
+        {
+            GenericLess<positive> cmp(*mapped);
+            APPLY_COMPARATOR(cmp)
+        }
+#undef DISPATCH_FOR_NULLABLE_COLUMN
+    }
+    else
+    {
+        GenericLess<positive> cmp(*mapped);
+        APPLY_COMPARATOR(cmp)
+    }
+#undef APPLY_COMPARATOR
 
     return ColumnArray::create(array.getData().permute(permutation, 0), array.getOffsetsPtr());
 }