### Changelog category (leave one):
- Performance Improvement


### Changelog entry (a user-readable short description of the changes that goes to CHANGELOG.md):
Optimize arraySort

### Documentation entry for user-facing changes

- [ ] Documentation is written (mandatory for new features)

<!---
Directly edit documentation source files in the "docs" folder with the same pull-request as code changes

or

Add a user-readable short description of the changes that should be added to docs.clickhouse.com below.

At a minimum, the following information should be added (but add more as needed).
- Motivation: Why is this function, table engine, etc. useful to ClickHouse users?

- Parameters: If the feature being added takes arguments, options or is influenced by settings, please list them below with a brief explanation.

- Example use: A query or command.
-->

``` sql
select arraySort(array(rand(), rand(), rand(), rand(), rand())) from numbers(100000000) format Null; 
Before
0 rows in set. Elapsed: 3.706 sec. Processed 100.00 million rows, 800.00 MB (26.98 million rows/s., 215.85 MB/s.)
0 rows in set. Elapsed: 3.903 sec. Processed 100.00 million rows, 800.00 MB (25.62 million rows/s., 204.94 MB/s.)
0 rows in set. Elapsed: 3.600 sec. Processed 100.00 million rows, 800.00 MB (27.78 million rows/s., 222.25 MB/s.)
After
0 rows in set. Elapsed: 2.846 sec. Processed 100.00 million rows, 800.00 MB (35.14 million rows/s., 281.14 MB/s.)
0 rows in set. Elapsed: 2.883 sec. Processed 100.00 million rows, 800.00 MB (34.68 million rows/s., 277.47 MB/s.)
0 rows in set. Elapsed: 2.827 sec. Processed 100.00 million rows, 800.00 MB (35.38 million rows/s., 283.03 MB/s.)

select arraySort(array(if(number = 0, rand(), null), rand(), rand(), rand(), rand())) from numbers(100000000) format Null; 
Before
0 rows in set. Elapsed: 7.064 sec. Processed 100.00 million rows, 800.00 MB (14.16 million rows/s., 113.24 MB/s.)
0 rows in set. Elapsed: 6.831 sec. Processed 100.00 million rows, 800.00 MB (14.64 million rows/s., 117.12 MB/s.)
0 rows in set. Elapsed: 6.871 sec. Processed 100.00 million rows, 800.00 MB (14.55 million rows/s., 116.43 MB/s.)
After
0 rows in set. Elapsed: 4.847 sec. Processed 100.00 million rows, 800.00 MB (20.63 million rows/s., 165.05 MB/s.)
0 rows in set. Elapsed: 4.929 sec. Processed 100.00 million rows, 800.00 MB (20.29 million rows/s., 162.30 MB/s.)
0 rows in set. Elapsed: 4.896 sec. Processed 100.00 million rows, 800.00 MB (20.42 million rows/s., 163.39 MB/s.)
```